refactor(service): service 패키지 생성 및 구조 변경
- LottoGenerator 클래스를 service 패키지로 이동
